backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.0
numpy==1.25.2
pydantic==2.3.0
torch==2.1.0
torchvision==0.16.0
facenet-pytorch==2.5.3
albumentations==1.3.1
efficientnet-pytorch==0.7.1
opencv-python==4.8.0.74
scikit-learn==1.3.0
grad-cam==1.4.8
scikit-image==0.21.0